Moving exported interfaces into interfaces.ts
