+++ Extension/src/LanguageServer/clientCollection.ts
@@ -8,7 +8,7 @@ import * as vscode from 'vscode';
 import * as util from '../common';
 import * as telemetry from '../telemetry';
 import * as cpptools from './client';
-import { Configuration } from './configurations';
+import { Configuration } from '../interfaces';
 
 const defaultClientKey: string = "@@default@@";
 export interface ClientKey {
